+++ blockfrost-api/CHANGELOG.md
@@ -6,6 +6,7 @@
   * `dataHash` field to `UtxoInput` and `UtxoOutput`
   * `outputIndex` field to `UtxoOutput`
   * `script` field to `AddressInfo`
+  * `/txs/${hash}/redeemers` endpoint with `TransactionRedeemer` and `ValidationPurpose` types
 
 # Version [0.1.0.0](https://github.com/blockfrost/blockfrost-haskell/compare/initial...0.1.0.0) (2021-09-14)